requirements.txt
pillow>=9.3.0
matplotlib~=3.6.2
numpy~=1.23.4